// src/features/order/schemas/get-order.schema.ts
import { z } from 'zod';

export const getOrderSchema = z.object({
  ordNbr: z.string(),
  ordDate: z.string(),
  cusId: z.string().nullish(),
  cusCode: z.string().nullish(),
  cusName: z.string().nullish(),
  address: z.string().nullish(),
  tel: z.string().nullish(),
  buyerIdentityCard: z.string().nullish(),
  email: z.string().nullish(),
  description: z.string().nullish(),
  totalAmountWithoutVat: z.number().nullish(),
  discountPercenatge: z.number().nullish(),
  totalDiscount: z.number().nullish(),
  taxAmount: z.number().nullish(),
  taxReductionAmount: z.number().nullish(),
  totalAmount: z.number().nullish(),
  cusAmount: z.number().nullish(),
  returnAmount: z.number().nullish(),
  paymentMethod: z.string().nullish(),
  orderStatus: z.string().nullish(),
  invoiceStatus: z.string().nullish(),
  invoiceNbr: z.string().nullish(),
  taxCode: z.string().nullish(),
  companyName: z.string().nullish(),
  securityNbr: z.string().nullish(),
  linkCheckInvoice: z.string().nullish(),
  invoiceType: z.string().nullish(),
  invoiceId: z.string().nullish(),
  invoiceDate: z.string().nullish(),
  invoiceSymbol: z.string().nullish(),
  totalVatAmount: z.number().nullish(),
  orderDetails: z.array(z.any()),
  bankId: z.string().nullish(),
  voucherCode: z.string().nullish(),
  warehouseId: z.string().nullish(),
  voucherOrd: z.string().nullish(),
  toWarehouseId: z.string().nullish(),
  isVat: z.boolean().nullish(),
  refId: z.string().nullish(),
  refNbr: z.string().nullish(),
  tenantId: z.string().nullish(),
  id: z.string(),
  creationTime: z.string().nullish(),
  creatorId: z.string().nullish(),
  lastModificationTime: z.string().nullish(),
  lastModifierId: z.string().nullish(),
});

export type GetOrderSchemaType = z.infer<typeof getOrderSchema>;